+++ clouddriver-openstack/src/main/groovy/com/netflix/spinnaker/clouddriver/openstack/client/OpenstackComputeProvider.groovy
@@ -83,10 +83,11 @@ interface OpenstackComputeProvider {
   /**
    * Looks up the port associated by vip and uses the deviceId to get the attached floatingIp.
    * @param region
+   * @param portDeviceId
    * @param vipId
    * @return
    */
-  FloatingIP getAssociatedFloatingIp(final String region, final String vipId)
+  FloatingIP getAssociatedFloatingIp(final String region, final String portDeviceId, final String vipId)
 
   /**
    * List all floating ips in the region.
